--- src/app/app.filter.ts
@@ -1,8 +1,0 @@
-export module Filtros {
-    export function cortarTexto() {
-        return function(input, limit){
-        return (input.length > limit && limit !== 0) ? input.substr(0, limit) + '...' : input;
-        };
-    }
-}
-
