+++ src/app/components/modal/modal.component.css
@@ -1,0 +1,27 @@
+.custom-modal-container {
+    padding: 15px;
+}
+
+.custom-modal-header[_ngcontent-c3] {
+    background-color: #000;
+    color: #fff;
+    -webkit-box-shadow: 0px 3px 5px 0px rgba(0,0,0,0.75);
+    -moz-box-shadow: 0px 3px 5px 0px rgba(0,0,0,0.75);
+    box-shadow: 0px 3px 5px 0px rgba(0,0,0,0.75);
+    margin-top: 0px;
+    margin-bottom: 0px;
+}
+
+.modal-content {
+    background-color: #000;
+    color: #fff;
+}
+.custom-modal-header {
+    background-color: #219161;
+    color: #fff;
+    -webkit-box-shadow: 0px 3px 5px 0px rgba(0,0,0,0.75);
+    -moz-box-shadow: 0px 3px 5px 0px rgba(0,0,0,0.75);
+    box-shadow: 0px 3px 5px 0px rgba(0,0,0,0.75);
+    margin-top: -15px;
+    margin-bottom: 40px;
+}
