+++ src/app/components/modal/modal.component.html
@@ -1,0 +1,8 @@
+<div class="container-fluid custom-modal-container">
+<div class="row custom-modal-header">
+    <div class="embed-responsive embed-responsive-16by9">
+    <iframe class="embed-responsive-item" frameborder="0" width="600" height="350" [src]="url"></iframe>
+    </div>
+    <div (click)="closeModal()"> cerrar</div><!--class="e2e-iframe-trusted-src"-->
+</div>
+</div>
